Fix the linker script so BSS goes in RAM
I'm surprised there are no warnings when a section like data or bss is
not defined in the linker script but they are used by the objects.

Anyway, in Convergence, we fastidiously placed mutable data in either
EWRAM or IWRAM as appropriate, and I guess we never used data or bss.

We adjust that here, putting our variables in IWRAM (it's a small game),
but also making the linker script more sane in general.