Import graphics conversion tools from Convergence
I apologize for the crude state of the OCaml here; circa 2002 this might
have been one of the first things I wrote in it, so there's a lot to be
improved.  However, because I actually want to finish this for January,
I might not get it refactored right now.  Caveat lector.